Move requesting logic into new file
